# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement Orange Money pour les transferts et paiements",
        "contact": {
            "email": "contact@om-pay.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://om-pay-api-2cst.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/payments/initiate": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Initier un paiement ou transfert",
                "operationId": "b61fb11d3bca59e8d0cce3f43f4db6f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "currency",
                                    "recipient_account"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant du paiement",
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "currency": {
                                        "type": "string",
                                        "enum": [
                                            "XOF",
                                            "EUR",
                                            "USD"
                                        ],
                                        "example": "XOF"
                                    },
                                    "recipient_account": {
                                        "type": "string",
                                        "example": "OM221772345678"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Paiement test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement initié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "integer"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "processed_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant"
                    },
                    "403": {
                        "description": "Compte non actif"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfers/initiate": {
            "post": {
                "tags": [
                    "Transferts"
                ],
                "summary": "Initier un transfert d'argent",
                "operationId": "41ac2725e62761b02b06bffe8d97a9ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "currency",
                                    "recipient_account"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant du transfert",
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "currency": {
                                        "type": "string",
                                        "enum": [
                                            "XOF",
                                            "EUR",
                                            "USD"
                                        ],
                                        "example": "XOF"
                                    },
                                    "recipient_account": {
                                        "type": "string",
                                        "example": "OM221772345678"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Transfert test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "integer"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "processed_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "recipient": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant"
                    },
                    "403": {
                        "description": "Compte non actif"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/status/{reference}": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Vérifier le statut d'une transaction",
                "operationId": "2c78952956fa0ee242116363f53fe6b5",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "path",
                        "description": "Référence de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut de la transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending",
                                                        "completed",
                                                        "failed",
                                                        "cancelled"
                                                    ]
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "debit",
                                                        "credit"
                                                    ]
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "processed_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}